cleanup - move code around
