+++ src/features/workspace/accept-invite.tsx
@@ -35,67 +35,16 @@ import { Pages } from "@/utils/pages.ts"
 import { CHECK_MAIL_REASON } from "@/constants.ts"
 import { useAcceptInvite } from "@/features/workspace/api/accept-invite.ts"
 import { useDebounce } from "@/hooks/use-debounce.ts"
-import { MIN_USERNAME_LENGTH, useCheckUsername } from "./api/check-username.ts"
+import { useCheckUsername } from "./api/check-username.ts"
 import {
 	InputGroup,
 	InputGroupAddon,
 	InputGroupInput,
 } from "@/components/ui/input-group"
-import { CircleCheckBig, CircleX, Loader2 } from "lucide-react"
-import { type AxiosError, HttpStatusCode } from "axios"
-import type { UseQueryResult } from "@tanstack/react-query"
+import UsernameStateIcon from "@/features/workspace/components/username-state-icon.tsx"
+import usernameCheckMessage from "@/features/workspace/utils/username-check-message.ts"
 
 const LAG_MS = 2500
-
-function usernameCheckMessage(
-	query: UseQueryResult<null, AxiosError<unknown, Error>>,
-	username: string,
-) {
-	if (username.length < MIN_USERNAME_LENGTH || !query.isError) return null
-
-	const status = query.error?.response?.status
-	if (status === HttpStatusCode.Conflict)
-		return "username taken, lets get creative!!"
-	if (status === HttpStatusCode.BadRequest)
-		return 'Invalid username pattern. Try something with pattern "john.doe" or just "john"'
-	return null
-}
-
-function UsernameStateIcon({
-	query,
-}: {
-	query: UseQueryResult<null, AxiosError<unknown, Error>>
-}) {
-	const icon = () => {
-		if (query.isSuccess) {
-			return (
-				<CircleCheckBig
-					data-testid="username-available"
-					className="size-4 text-green-600"
-				/>
-			)
-		}
-		if (
-			query.isError &&
-			(query.error?.response?.status === HttpStatusCode.Conflict ||
-				query.error?.response?.status === HttpStatusCode.BadRequest)
-		) {
-			return (
-				<CircleX
-					data-testid="username-error"
-					className="size-4 text-destructive"
-				/>
-			)
-		}
-		return (
-			<Loader2
-				data-testid="username-checking"
-				className="size-4 animate-spin text-neutral-400"
-			/>
-		)
-	}
-	return icon()
-}
 
 export function AcceptInvite() {
 	const { inviteCode } = useSearch({ from: "/workspace-invite" })
